updated for Visual Studio Express, various other additions and fixes

+++ EZ-USB/RX-FFT/LibRXFFT/Libraries/SignalProcessing/AudioDemodulator.cs
@@ -98,7 +98,7 @@ namespace LibRXFFT.Libraries.SignalProcessing
             }
         }
 
-        private void SetSamplingRate(long rate)
+        private void SetSamplingRate(double rate)
         {
             lock (DemodState.SoundSinkInfos)
             {
